requirements.txt
django==5.2.4
djangorestframework==3.16.1
mysqlclient==2.2.7
django-filter==24.3